Attaching event listeners to groups of nodes
This commit adds the exercises from The Odin Project DOM manipulation
and events tutorial that have to do with attaching listeners to multiple
document object model element nodes without have to repeat lines of JS
code.

This commit also adds the e, i.e., event, parameter to a function call.
This is an event object which references the event itself that triggered
the listener and subsequent function call.